Separate files due to syntax incompatibilities between PHP 8.4 and 8.1
