+++ tests/IntegrationTest/Definitions/CreateDefinitionTest.php
@@ -25,6 +25,15 @@ use DI\Definition\Exception\InvalidDefinition;
  */
 class CreateDefinitionTest extends BaseContainerTest
 {
+    public static function setUpBeforeClass(): void
+    {
+        if (PHP_VERSION_ID < 80400) {
+            require_once __DIR__ . '/CreateDefinition/class-php83.php';
+        } else {
+            require_once __DIR__ . '/CreateDefinition/class.php';
+        }
+    }
+
     /**
      * @dataProvider provideContainer
      */
@@ -307,106 +316,5 @@ class CreateDefinitionTest extends BaseContainerTest
 
         self::assertEntryIsCompiled($container, 'foo');
         self::assertInstanceOf(Property::class, $container->get('foo')['bar']);
-    }
-}
-
-namespace DI\Test\IntegrationTest\Definitions\CreateDefinitionTest;
-
-class Property
-{
-    public $foo;
-}
-
-class ConstructorInjection
-{
-    public $value;
-    public $scalarValue;
-    public $typedValue;
-    public $typedOptionalValue;
-    public $typedOptionalValueDefaultNull;
-    /** @var \ProxyManager\Proxy\LazyLoadingInterface */
-    public $lazyService;
-    public $optionalValue;
-
-    public function __construct(
-        $value,
-        string $scalarValue,
-        \stdClass $typedValue,
-        \stdClass $lazyService,
-        ?\stdClass $typedOptionalValue = new \stdClass(),
-        ?\stdClass $typedOptionalValueDefaultNull = null,
-        $optionalValue = 'hello'
-    ) {
-        $this->value = $value;
-        $this->scalarValue = $scalarValue;
-        $this->typedValue = $typedValue;
-        $this->typedOptionalValue = $typedOptionalValue;
-        $this->typedOptionalValueDefaultNull = $typedOptionalValueDefaultNull;
-        $this->lazyService = $lazyService;
-        $this->optionalValue = $optionalValue;
-    }
-}
-
-class PropertyInjection
-{
-    public $value;
-    public $entry;
-    /** @var \ProxyManager\Proxy\LazyLoadingInterface */
-    public $lazyService;
-}
-
-class MethodInjection
-{
-    public $value;
-    public $scalarValue;
-    public $typedValue;
-    public $typedOptionalValue;
-    public $typedOptionalValueDefaultNull;
-    /** @var \ProxyManager\Proxy\LazyLoadingInterface */
-    public $lazyService;
-    public $optionalValue;
-
-    public function method(
-        $value,
-        string $scalarValue,
-        \stdClass $typedValue,
-        \stdClass $lazyService,
-        ?\stdClass $typedOptionalValue = new stdClass(),
-        ?\stdClass $typedOptionalValueDefaultNull = null,
-        $optionalValue = 'hello'
-    ) {
-        $this->value = $value;
-        $this->scalarValue = $scalarValue;
-        $this->typedValue = $typedValue;
-        $this->typedOptionalValue = $typedOptionalValue;
-        $this->typedOptionalValueDefaultNull = $typedOptionalValueDefaultNull;
-        $this->lazyService = $lazyService;
-        $this->optionalValue = $optionalValue;
-    }
-}
-
-class PrivatePropertyInjection
-{
-    private $private;
-    protected $protected;
-
-    public function getPrivate()
-    {
-        return $this->private;
-    }
-
-    public function getProtected()
-    {
-        return $this->protected;
-    }
-}
-
-class PrivatePropertyInjectionSubClass extends PrivatePropertyInjection
-{
-    private $private;
-
-    public function getSubClassPrivate()
-    {
-        return $this->private;
     }
 }
